ColorManagement: Moved ColorSpace and TransferFunctions logic into into Core module.
This is needed in order to use the working color space during the luminance calculation of any color, and avoid circular dependencies between the ColorManagement and Core modules. Header deprecation and module dependency removal will be done in a subsequent CL.

#jira UE-201221
#rb Sebastien.Hillaire
[FYI] rod.bogart, matt.peters

[CL 32299564 by eric renaudhoude in ue5-main branch]

+++ Engine/Source/Runtime/Core/Private/ColorManagement/ColorSpace.cpp
@@ -1,6 +1,6 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
-#include "ColorSpace.h"
+#include "ColorManagement/ColorSpace.h"
 #include "Misc/ConfigCacheIni.h"
 #include "Math/VectorRegister.h"
 
@@ -246,7 +246,7 @@ FLinearColor FColorSpace::MakeFromColorTemperature(float Temp) const
 float FColorSpace::GetLuminance(const FLinearColor& Color) const
 {
 	//Note: Equivalent to the dot product of Color and RgbToXYZ.GetColumn(1).
-	return Color.R * RgbToXYZ.M[0][1] + Color.G * RgbToXYZ.M[1][1] + Color.B * RgbToXYZ.M[2][1];
+	return Color.R * (float)RgbToXYZ.M[0][1] + Color.G * (float)RgbToXYZ.M[1][1] + Color.B * (float)RgbToXYZ.M[2][1];
 }
 
 FLinearColor FColorSpace::GetLuminanceFactors() const
